Remove unused reducer and move tests
